calculations fixed, vin api fixed

+++ truckCostTracker-laravel/database/migrations/2019_08_19_000000_create_failed_jobs_table.php
@@ -13,11 +13,12 @@ return new class extends Migration
      */
     public function up()
     {
+        // This is basically where I defined the schema and the requirements for the model. Its the Blue Print!
         Schema::create('trips', function (Blueprint $table) {
             $table->id();
             $table->string('trip_name')->unique();
             $table->string('vin_number');
-            $table->integer('distance');
+            $table->float('distance');
             $table->integer('tolls');
             $table->integer('foodBudget');
             $table->integer('wearAndTear');
